+++ tests/test_jogo.py
@@ -10,7 +10,7 @@ from jogo import (
     serializar_mapa,
     verificar_level_up,
 )
-from src.entidades import Inimigo, Item, Personagem, Sala
+from src.entidades import Inimigo, Item, Moeda, Personagem, Sala
 
 
 # Fixture para criar um jogador base para os testes
@@ -33,6 +33,7 @@ def jogador_base() -> Personagem:
         nivel=1,
         xp_atual=0,
         xp_para_proximo_nivel=100,
+        carteira=Moeda.from_gp_sp_cp(),
     )